requirements.txt
# =============================================================================
# SakaiBot Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# For development: pip install -r requirements.txt && pip install -e ".[dev]"
# =============================================================================

# -----------------------------------------------------------------------------
# Core - Telegram Client
# -----------------------------------------------------------------------------
telethon>=1.34.0
cryptg>=0.4.0        # Faster cryptography for Telethon

# -----------------------------------------------------------------------------
# AI and LLM Providers
# -----------------------------------------------------------------------------
openai>=1.0.0                # OpenRouter API (OpenAI-compatible)
google-genai>=0.8.0          # Gemini API (unified SDK for all features)
google-generativeai>=0.8.0   # Gemini API (legacy SDK - required for google.generativeai imports)

# -----------------------------------------------------------------------------
# Speech Processing (STT/TTS)
# -----------------------------------------------------------------------------
SpeechRecognition>=3.10.0    # Speech-to-text
pydub>=0.25.1                # Audio processing
gTTS>=2.5.1                  # Google Text-to-Speech (fallback)

# Note: azure-cognitiveservices-speech removed - optional and heavy dependency
# If needed, install separately: pip install azure-cognitiveservices-speech

# -----------------------------------------------------------------------------
# Configuration and Validation
# -----------------------------------------------------------------------------
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# -----------------------------------------------------------------------------
# CLI and UI
# -----------------------------------------------------------------------------
click>=8.1.0          # Command-line interface
rich>=13.0.0          # Rich terminal output
tabulate>=0.9.0       # Table formatting

# -----------------------------------------------------------------------------
# Async Utilities
# -----------------------------------------------------------------------------
aiofiles>=23.0.0      # Async file operations
aiohttp>=3.9.0        # Async HTTP client (for image generation workers)
httpx[socks]>=0.25.0  # HTTP client with async and SOCKS proxy support

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
pytz>=2023.3          # Timezone handling
psutil>=5.9.0         # System resource monitoring (CPU, memory)
# -----------------------------------------------------------------------------
# Testing (optional - for development)
# -----------------------------------------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0
freezegun>=1.2.0
